reorganize test hierarchy; closes #2838
