+++ Makefile
@@ -63,79 +63,79 @@ test-jsapi:
 
 test-unit:
 	@printf "==> [Test :: Unit]\n"
-	$(call test_node,unit) test/acceptance/*.js \
-		--growl \
-		test/*.js
+	$(call test_node,unit) test/unit/*.spec.js \
+		test/node-unit/*.spec.js \
+		--growl
 
 test-integration:
 	@printf "==> [Test :: Integrations]\n"
 	$(call test_node,integration) --timeout 5000 \
-		test/integration/*.js
+		test/integration/*.spec.js
 
 test-compilers:
 	@printf "==> [Test :: Compilers]\n"
 	$(call test_node,compilers) --compilers coffee:coffee-script/register,foo:./test/compiler/foo \
-		test/acceptance/test.coffee \
-		test/acceptance/test.foo
+		test/compiler/test.coffee \
+		test/compiler/test.foo
 
 test-requires:
 	@printf "==> [Test :: Requires]\n"
 	$(call test_node,requires) --compilers coffee:coffee-script/register \
-		--require test/acceptance/require/a.js \
-		--require test/acceptance/require/b.coffee \
-		--require test/acceptance/require/c.js \
-		--require test/acceptance/require/d.coffee \
-		test/acceptance/require/require.spec.js
+		--require test/require/a.js \
+		--require test/require/b.coffee \
+		--require test/require/c.js \
+		--require test/require/d.coffee \
+		test/require/require.spec.js
 
 test-bdd:
 	@printf "==> [Test :: BDD]\n"
 	$(call test_node,bdd) --ui bdd \
-		test/acceptance/interfaces/bdd.spec
+		test/interfaces/bdd.spec
 
 test-tdd:
 	@printf "==> [Test :: TDD]\n"
 	$(call test_node,tdd) --ui tdd \
-		test/acceptance/interfaces/tdd.spec
+		test/interfaces/tdd.spec
 
 test-qunit:
 	@printf "==> [Test :: QUnit]\n"
 	$(call test_node,qunit) --ui qunit \
-		test/acceptance/interfaces/qunit.spec
+		test/interfaces/qunit.spec
 
 test-exports:
 	@printf "==> [Test :: Exports]\n"
 	$(call test_node,exports) --ui exports \
-		test/acceptance/interfaces/exports.spec
+		test/interfaces/exports.spec
 
 test-glob:
 	@printf "==> [Test :: Glob]\n"
-	bash ./test/acceptance/glob/glob.sh
+	bash ./test/glob/glob.sh
 
 test-reporters:
 	@printf "==> [Test :: Reporters]\n"
-	$(call test_node,reporters) test/reporters/*.js
+	$(call test_node,reporters) test/reporters/*.spec.js
 
 test-only:
 	@printf "==> [Test :: Only]\n"
 	$(call test_node,only-tdd) --ui tdd \
-		test/acceptance/misc/only/tdd.spec
+		test/misc/only/tdd.spec
 
 	$(call test_node,only-bdd) --ui bdd \
-		test/acceptance/misc/only/bdd.spec
+		test/misc/only/bdd.spec
 
 	$(call test_node,only-bdd-require) --ui qunit \
-		test/acceptance/misc/only/bdd-require.spec
+		test/misc/only/bdd-require.spec
 
 test-global-only:
 	@printf "==> [Test :: Global Only]\n"
 	$(call test_node,global-only-tdd) --ui tdd \
-		test/acceptance/misc/only/global/tdd.spec
+		test/misc/only/global/tdd.spec
 
 	$(call test_node,global-only-bdd) --ui bdd \
-		test/acceptance/misc/only/global/bdd.spec
+		test/misc/only/global/bdd.spec
 
 	$(call test_node,global-only-qunit) --ui qunit \
-		test/acceptance/misc/only/global/qunit.spec
+		test/misc/only/global/qunit.spec
 
 test-mocha:
 	@printf "==> [Test :: Mocha]\n"
@@ -144,19 +144,19 @@ test-mocha:
 non-tty:
 	@printf "==> [Test :: Non-TTY]\n"
 	$(call test_node,non-tty-dot) --reporter dot \
-		test/acceptance/interfaces/bdd.spec 2>&1 > /tmp/dot.out
+		test/interfaces/bdd.spec 2>&1 > /tmp/dot.out
 
 	@echo dot:
 	@cat /tmp/dot.out
 
 	$(call test_node,non-tty-list) --reporter list \
-		test/acceptance/interfaces/bdd.spec 2>&1 > /tmp/list.out
+		test/interfaces/bdd.spec 2>&1 > /tmp/list.out
 
 	@echo list:
 	@cat /tmp/list.out
 
 	$(call test_node,non-tty-spec) --reporter spec \
-		test/acceptance/interfaces/bdd.spec 2>&1 > /tmp/spec.out
+		test/interfaces/bdd.spec 2>&1 > /tmp/spec.out
 
 	@echo spec:
 	@cat /tmp/spec.out
