+++ test/browser-fixtures/exports.fixture.js
@@ -2,5 +2,7 @@
 
 /* eslint-env browser */
 
+process.stdout = require('browser-stdout')();
+
 window.mocha.timeout(200)
   .ui('exports');
