+++ test/browser/README.md
@@ -1,0 +1,3 @@
+These files need to be run manually by loading the `.html` file(s) in a browser.
+
+It would be awesome if we could automate that!
