+++ test/compiler/test.foo
