+++ test/glob/glob.spec.js
