+++ test/jsapi/index.js
@@ -12,16 +12,16 @@ var mocha = new Mocha({
 // mocha.reporter('spec');
 require('should');
 
-mocha.addFile('test/suite.spec.js');
-mocha.addFile('test/runner.spec.js');
-mocha.addFile('test/runnable.spec.js');
-mocha.addFile('test/hook-sync.spec.js');
-mocha.addFile('test/hook-sync-nested.spec.js');
-mocha.addFile('test/hook-async.spec.js');
-mocha.addFile('test/acceptance/duration.spec.js');
-mocha.addFile('test/acceptance/fs.spec.js');
-mocha.addFile('test/acceptance/globals.spec.js');
-mocha.addFile('test/acceptance/timeout.spec.js');
+mocha.addFile('test/unit/suite.spec.js');
+mocha.addFile('test/unit/runner.spec.js');
+mocha.addFile('test/unit/runnable.spec.js');
+mocha.addFile('test/unit/hook-sync.spec.js');
+mocha.addFile('test/unit/hook-sync-nested.spec.js');
+mocha.addFile('test/unit/hook-async.spec.js');
+mocha.addFile('test/unit/duration.spec.js');
+mocha.addFile('test/node-unit/fs.spec.js');
+mocha.addFile('test/unit/globals.spec.js');
+mocha.addFile('test/unit/timeout.spec.js');
 
 mocha.run(function () {
   console.log('done');
