+++ test/misc/exit.spec.js
