+++ test/misc/nontty.spec.js
