+++ test/node-unit/color.spec.js
