+++ test/node-unit/file-utils.spec.js
