+++ test/node-unit/fs.spec.js
