+++ test/require/a.js
