+++ test/require/b.coffee
