+++ test/unit/duration.spec.js
