+++ test/unit/globals.spec.js
