+++ test/unit/hook-async.spec.js
