+++ test/unit/hook-sync-nested.spec.js
