+++ test/unit/hook-sync.spec.js
