+++ test/unit/hook-timeout.spec.js
