+++ test/unit/root.spec.js
