+++ test/unit/runnable.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var mocha = require('../');
+var mocha = require('../../lib/mocha');
 var utils = mocha.utils;
 var Runnable = mocha.Runnable;
 
@@ -183,7 +183,9 @@ describe('Runnable(title, fn)', function () {
     describe('when timeouts are disabled', function () {
       it('should not error with timeout', function (done) {
         var test = new Runnable('foo', function (done) {
-          setTimeout(process.nextTick.bind(undefined, done), 2);
+          setTimeout(function () {
+            setTimeout(done);
+          }, 2);
         });
         test.timeout(1);
         test.enableTimeouts(false);
@@ -195,7 +197,7 @@ describe('Runnable(title, fn)', function () {
       describe('without error', function () {
         it('should invoke the callback', function (done) {
           var test = new Runnable('foo', function (done) {
-            process.nextTick(done);
+            setTimeout(done);
           });
 
           test.run(done);
@@ -210,9 +212,9 @@ describe('Runnable(title, fn)', function () {
 
             var test = new Runnable('foo', function (done) {
               process.nextTick(done);
-              process.nextTick(done);
-              process.nextTick(done);
-              process.nextTick(done);
+              setTimeout(done);
+              setTimeout(done);
+              setTimeout(done);
             });
 
             test.on('error', function (err) {
@@ -236,10 +238,10 @@ describe('Runnable(title, fn)', function () {
 
             var test = new Runnable('foo', function (done) {
               done(new Error('fail'));
-              process.nextTick(done);
+              setTimeout(done);
               done(new Error('fail'));
-              process.nextTick(done);
-              process.nextTick(done);
+              setTimeout(done);
+              setTimeout(done);
             });
 
             test.on('error', function (err) {
@@ -359,7 +361,7 @@ describe('Runnable(title, fn)', function () {
       describe('when the promise is fulfilled with no value', function () {
         var fulfilledPromise = {
           then: function (fulfilled, rejected) {
-            process.nextTick(fulfilled);
+            setTimeout(fulfilled);
           }
         };
 
@@ -375,7 +377,7 @@ describe('Runnable(title, fn)', function () {
       describe('when the promise is fulfilled with a value', function () {
         var fulfilledPromise = {
           then: function (fulfilled, rejected) {
-            process.nextTick(function () {
+            setTimeout(function () {
               fulfilled({});
             });
           }
@@ -394,7 +396,7 @@ describe('Runnable(title, fn)', function () {
         var expectedErr = new Error('fail');
         var rejectedPromise = {
           then: function (fulfilled, rejected) {
-            process.nextTick(function () {
+            setTimeout(function () {
               rejected(expectedErr);
             });
           }
@@ -416,7 +418,7 @@ describe('Runnable(title, fn)', function () {
         var expectedErr = new Error('Promise rejected with no or falsy reason');
         var rejectedPromise = {
           then: function (fulfilled, rejected) {
-            process.nextTick(function () {
+            setTimeout(function () {
               rejected();
             });
           }
