+++ test/unit/throw.spec.js
